experimental bundle valuation distribution; name changes

+++ src/main/java/brown/platform/accounting/IAccount.java
@@ -16,7 +16,7 @@ public interface IAccount {
   /**
    * @return ID of agent owning the account
    */
-  public double getID();
+  public int getID();
 
   /**
    * @return double representing amount of money in account
